Flask-API/requirements.txt
blinker==1.8.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
Flask-Cors==5.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.1
Werkzeug==3.0.4
dtw-python == 1.5.3
numpy==1.26.4
pandas==2.1.3
statsmodels==0.14.4
scipy==1.15.3